Fix/report sending (#190)
* feature/add_branch_to_build_tools

* feature/add_branch_to_build_tools

* feature/add_branch_to_build_tools

* feature/add_branch_to_build_tools

* feature/add_branch_to_build_tools

* feature/add_branch_to_build_tools

* feature/add_branch_to_build_tools

* feature/add_branch_to_build_tools

* feature/add_branch_to_build_tools

* feature/add_branch_to_build_tools

* feature/add_branch_to_build_tools

* feature/add_branch_to_build_tools

* feature/add_branch_to_build_tools

* feature/add_branch_to_build_tools

* feature/add_branch_to_build_tools

* feature/add_branch_to_build_tools

* feature/add_branch_to_build_tools

* feature/add_branch_to_build_tools

* feature/add_branch_to_build_tools

* Refactoring (#189)

* refactoring

* refactoring

---------


* refactoring

* refactoring

* refactoring

---------

+++ tasks.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from os import getcwd, system
-from os.path import join
+from os.path import join, isfile
 from host_tools import Process, Service
 from elevate import elevate
 from host_tools.utils import Dir
@@ -34,7 +34,8 @@ def desktop_test(
         appimage: bool = False,
         flatpak: bool = False,
         open_retries: int = None,
-        retest: bool = False
+        retest: bool = False,
+        only_portal: bool = False
 ):
     num_processes = int(processes) if processes else 1
 
@@ -51,18 +52,24 @@ def desktop_test(
         retest=retest
     )
 
-    if num_processes > 1 and not name and len(data.vm_names) > 1:
-        data.status_bar = False
-        multiprocess.run(DesktopTest, data, num_processes, 10, headless)
-    else:
-        data.status_bar = True
-        for vm in Vbox().check_vm_names([name] if name else data.vm_names):
-            DesktopTest(vm, data).run(headless=headless)
-
     report = DesktopReport(report_path=data.full_report_path)
-    report.get_full(data.version)
-    report.send_to_tg(data=data) if not name else None
-    report.send_to_report_portal(data.portal_project_name, data.package_name) if connect_portal else None
+
+    if not only_portal:
+        if num_processes > 1 and not name and len(data.vm_names) > 1:
+            data.status_bar = False
+            multiprocess.run(DesktopTest, data, num_processes, 10, headless)
+        else:
+            data.status_bar = True
+            for vm in Vbox().check_vm_names([name] if name else data.vm_names):
+                DesktopTest(vm, data).run(headless=headless)
+
+        report.get_full(data.version)
+
+    if only_portal and not isfile(data.full_report_path):
+        raise FileNotFoundError(f"Report file {data.full_report_path} not found")
+
+    report.send_to_tg(data=data) if not name and not only_portal else None
+    report.send_to_report_portal(data.portal_project_name, data.package_name) if connect_portal or only_portal else None
 
     error_vms = report.get_error_vm_list()
     if len(error_vms) > 0:
@@ -79,30 +86,35 @@ def builder_test(
         headless: bool = False,
         connect_portal: bool = False,
         telegram: bool = False,
+        only_portal: bool = False
 ):
     num_processes = int(processes) if processes else 1
     data = BuilderTestData(
         version=version or Prompt.ask('[red]Please enter version'),
         config_path=join(getcwd(), "builder_tests_config.json")
     )
+    if not only_portal:
+        builder = DocBuilder(version=data.version)
+        builder.get(dep_test_branch=data.dep_test_branch, builder_samples_branch=data.document_builder_samples)
+        builder.compress_dep_tests(delete=False)
+        Dir.delete(builder.local_path.dep_test_path)
 
-    builder = DocBuilder(version=data.version)
-    builder.get(dep_test_branch=data.dep_test_branch, builder_samples_branch=data.document_builder_samples)
-    builder.compress_dep_tests(delete=False)
-    Dir.delete(builder.local_path.dep_test_path)
+        if num_processes > 1 and not name and len(data.vm_names) > 1:
+            data.status_bar = False
+            multiprocess.run(BuilderTests, data, num_processes, 10, headless)
+        else:
+            data.status_bar = True
+            for vm in Vbox().check_vm_names([name] if name else data.vm_names):
+                BuilderTests(vm, data).run(headless=headless)
 
-    if num_processes > 1 and not name and len(data.vm_names) > 1:
-        data.status_bar = False
-        multiprocess.run(BuilderTests, data, num_processes, 10, headless)
-    else:
-        data.status_bar = True
-        for vm in Vbox().check_vm_names([name] if name else data.vm_names):
-            BuilderTests(vm, data).run(headless=headless)
+        data.report.get_full(data.version)
 
-    data.report.get_full(data.version)
+    if only_portal and not isfile(data.full_report_path):
+        raise FileNotFoundError(f"Report file {data.full_report_path} not found")
+
     report_sender = BuilderReportSender(report_path=data.report.path)
-    report_sender.to_telegram() if telegram else None
-    report_sender.to_report_portal(project_name=data.portal_project_name) if connect_portal else None
+    report_sender.to_telegram() if telegram and not only_portal else None
+    report_sender.to_report_portal(project_name=data.portal_project_name) if connect_portal or only_portal else None
 
 
 @task
